requirements.txt
anyio==3.6.1
motor==3.0.0
black==22.10.0
bson==0.5.10
click==8.1.3
fastapi==0.85.0
flake8==5.0.4
h11==0.14.0
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.18.3
pymongo==4.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5